HBX-1910: Refactor JdbcMetadataBuilder into smaller responsibilities - Extract JdbcMetadataBuilder#bindRootClass(...) into new class 'org.hibernate.tool.internal.reveng.binder.RootClassBinder'
